+++ genno/compat/xarray.py
@@ -1,6 +1,7 @@
 """Compatibility with :mod:`xarray`."""
 from typing import (
     Any,
+    Callable,
     Dict,
     Generic,
     Hashable,
@@ -36,74 +37,84 @@ class DataArrayLike(Generic[T]):
     __slots__: Tuple[str, ...] = tuple()
 
     # Type hints for mypy in downstream applications
-    # @abstractmethod  # NB this suppresses a mypy "empty-body" error
     def __len__(self) -> int:
-        return NotImplemented  # pragma: no cover
+        return NotImplemented
 
-    # @abstractmethod  # NB this suppresses a mypy "empty-body" error
     def __mul__(self, other):  # TODO set the return type
-        ...  # pragma: no cover
+        ...
 
-    # @abstractmethod
     def __pow__(self, other):  # TODO set the return type
-        ...  # pragma: no cover
+        ...
 
     def __radd__(self, other):
-        ...  # pragma: no cover
+        ...
 
     def __rmul__(self, other):
-        ...  # pragma: no cover
+        ...
 
     def __rsub__(self, other):
-        ...  # pragma: no cover
+        ...
 
     def __rtruediv__(self, other):
-        ...  # pragma: no cover
+        ...
 
-    # @abstractmethod  # NB this suppresses a mypy "empty-body" error
     def __truediv__(self, other):  # TODO set the return type
-        ...  # pragma: no cover
+        ...
+
+    def __neg__(self):
+        ...
 
     @property
     def attrs(self) -> Dict[Any, Any]:
-        return NotImplemented  # pragma: no cover
+        return NotImplemented
 
     @property
-    # @abstractmethod
+    def data(self) -> Any:
+        return NotImplemented
+
+    @property
     def coords(
         self,
     ) -> xarray.core.coordinates.DataArrayCoordinates:
-        return NotImplemented  # pragma: no cover
+        return NotImplemented
 
     @property
-    # @abstractmethod
     def dims(self) -> Tuple[Hashable, ...]:
-        return NotImplemented  # pragma: no cover
+        return NotImplemented
 
     def assign_coords(
         self,
         coords: Optional[Mapping[Any, Any]] = None,
         **coords_kwargs: Any,
     ):
-        ...  # pragma: no cover
+        ...
 
-    # @abstractmethod
+    def astype(
+        self,
+        dtype,
+        *,
+        order=None,
+        casting=None,
+        subok=None,
+        copy=None,
+        keep_attrs=True,
+    ):
+        ...
+
     def bfill(
         self,
         dim: Hashable,
         limit: Optional[int] = None,
     ):  # TODO set the return type
-        ...  # pragma: no cover
+        ...
 
-    # @abstractmethod
     def copy(
         self,
         deep: bool = True,
         data: Any = None,
     ):  # TODO set the return type
-        ...  # pragma: no cover
+        ...
 
-    # @abstractmethod
     def cumprod(
         self,
         dim: Dims = None,
@@ -112,7 +123,7 @@ class DataArrayLike(Generic[T]):
         keep_attrs: Optional[bool] = None,
         **kwargs: Any,
     ):  # TODO set the return type
-        ...  # pragma: no cover
+        ...
 
     def drop_vars(
         self,
@@ -120,7 +131,7 @@ class DataArrayLike(Generic[T]):
         *,
         errors="raise",
     ):
-        ...  # pragma: no cover
+        ...
 
     def expand_dims(
         self,
@@ -128,15 +139,14 @@ class DataArrayLike(Generic[T]):
         axis=None,
         **dim_kwargs: Any,
     ):  # TODO set the return type
-        ...  # pragma: no cover
+        ...
 
-    # @abstractmethod
     def ffill(
         self,
         dim: Hashable,
         limit: Optional[int] = None,
     ):  # TODO set the return type
-        return NotImplemented  # pragma: no cover
+        return NotImplemented
 
     def groupby(
         self,
@@ -144,7 +154,7 @@ class DataArrayLike(Generic[T]):
         squeeze: bool = True,
         restore_coord_dims: bool = False,
     ):
-        ...  # pragma: no cover
+        ...
 
     def interp(
         self,
@@ -154,21 +164,28 @@ class DataArrayLike(Generic[T]):
         kwargs: Optional[Mapping[str, Any]] = None,
         **coords_kwargs: Any,
     ):
-        ...  # pragma: no cover
+        ...
 
     def item(self, *args):
-        ...  # pragma: no cover
+        ...
 
-    # @abstractmethod
+    def pipe(
+        self,
+        func: Union[Callable[..., T], Tuple[Callable[..., T], str]],
+        *args: Any,
+        **kwargs: Any,
+    ) -> T:
+        return NotImplemented
+
     def rename(
         self,
         new_name_or_name_dict: Union[Hashable, Mapping[Any, Hashable]] = None,
         **names: Hashable,
     ):
-        ...  # pragma: no cover
+        ...
 
     def round(self, *args, **kwargs):
-        ...  # pragma: no cover
+        ...
 
     def sel(
         self,
@@ -177,16 +194,16 @@ class DataArrayLike(Generic[T]):
         tolerance=None,
         drop: bool = False,
         **indexers_kwargs: Any,
-    ):  # TODO set the return type
-        return NotImplemented  # pragma: no cover
+    ) -> T:
+        return NotImplemented
 
     def shift(
         self,
-        shifts: Optional[Mapping[Hashable, int]] = None,
+        shifts: Optional[Mapping[Any, int]] = None,
         fill_value: Any = None,
         **shifts_kwargs: int,
-    ):  # TODO set the return type
-        ...  # pragma: no cover
+    ):
+        ...
 
     def sum(
         self,
@@ -197,21 +214,21 @@ class DataArrayLike(Generic[T]):
         min_count: Optional[int] = None,
         keep_attrs: Optional[bool] = None,
         **kwargs: Any,
-    ):  # TODO set the return type
-        ...  # pragma: no cover
+    ):
+        ...
 
     def to_dataframe(
         self,
         name: Optional[Hashable] = None,
         dim_order: Optional[Sequence[Hashable]] = None,
     ) -> pd.DataFrame:
-        ...  # pragma: no cover
+        ...
 
     def to_numpy(self) -> np.ndarray:
-        return NotImplemented  # pragma: no cover
+        return NotImplemented
 
     def to_series(self) -> pd.Series:
         """Like :meth:`xarray.DataArray.to_series`."""
         # Provided only for type-checking in other packages. AttrSeries implements;
         # SparseDataArray uses the xr.DataArray method.
-        return NotImplemented  # pragma: no cover
+        return NotImplemented